Reorganize markdown-it directory
