+++ packages/core/src/lib/markdown-it/patches/markdown-it-emoji-fixed.js
