+++ packages/core/src/lib/markdown-it/patches/markdown-it-escape-special-tags.js
@@ -1,4 +1,4 @@
-const markdownIt = require('./index');
+const markdownIt = require('../index');
 
 /*
  Custom patch for the api to escape content in certain special tags
