+++ packages/core/src/lib/markdown-it/plugins/markdown-it-block-embed/PluginEnvironment.js
