+++ packages/core/src/lib/markdown-it/plugins/markdown-it-icons.js
