+++ packages/core/src/patches/index.js
@@ -1,5 +1,5 @@
 const htmlparser2patch = require('./htmlparser2');
-const markdownItEscapeSpecialTags = require('../lib/markdown-it/markdown-it-escape-special-tags');
+const markdownItEscapeSpecialTags = require('../lib/markdown-it/patches/markdown-it-escape-special-tags');
 
 function ignoreTags(tagsToIgnore) {
   htmlparser2patch.injectIgnoreTags(tagsToIgnore);
